backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# Pin bcrypt to 4.0.1 — newer 4.1+ removes `__about__` which passlib 1.7.4 reads
bcrypt==4.0.1
python-multipart==0.0.12
httpx==0.27.2
python-dateutil==2.9.0
